requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
asyncpg==0.29.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dill==0.3.8
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
inquirerpy==0.3.4
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
multiprocess==0.70.14
nest-asyncio==1.6.0
nestd==0.3.1
orjson==3.9.15
packaging==24.1
passlib==1.7.4
pfzy==0.3.4
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic-bin @ git+https://github.com/julian-r/python-magic.git@6029e2d43ce0ee9f268c1f112c70e5417493190f
python-multipart==0.0.9
pytz==2024.1
requests==2.32.3
robyn==0.58.1
rustimport==1.3.4
six==1.16.0
SQLAlchemy==2.0.31
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.2
uuid==1.30
uvloop==0.19.0
watchdog==2.2.1
wcwidth==0.2.13
whitenoise==6.7.0
